backend-python-app/requirements.txt
fastapi
uvicorn
hf_xet
python-dotenv
accelerate
pyjwt
pyngrok
